+++ R/imports.R
@@ -1,6 +1,7 @@
 
 #' @importFrom utils available.packages citation contrib.url download.file
-#'   file.edit getCRANmirrors head install.packages installed.packages
-#'   old.packages packageDescription packageVersion read.table remove.packages
-#'   sessionInfo str tail tar toBibtex untar update.packages unzip URLencode zip
+#'   download.packages file.edit getCRANmirrors head install.packages
+#'   installed.packages old.packages packageDescription packageVersion
+#'   read.table remove.packages sessionInfo str tail tar toBibtex untar
+#'   update.packages unzip URLencode zip
 NULL