+++ R/init.R
@@ -119,7 +119,7 @@ init <- function(project = NULL,
   if (action == "init") {
     vwritef("* Initializing project ...")
     renv_libpaths_activate(project = project)
-    renv_bootstrap_impl(project)
+    renv_imbue_impl(project)
     hydrate(project = project, library = library)
     snapshot(project = project, library = library, confirm = FALSE)
   } else if (action == "restore") {