+++ R/migrate.R
@@ -91,7 +91,7 @@ renv_migrate_packrat <- function(project = NULL, components = NULL) {
     callback(project)
 
   renv_migrate_packrat_infrastructure(project)
-  renv_bootstrap_impl(project)
+  renv_imbue_impl(project)
 
   fmt <- "* Project '%s' has been migrated from Packrat to renv."
   vwritef(fmt, aliased_path(project))